josekifix: move override loading logic to josekifixload.c
