+++ josekifix/josekifixload.c
@@ -1,17 +1,498 @@
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <math.h>
 
 #define DEBUG
 #include "board.h"
 #include "debug.h"
+#include "version.h"
 #include "engine.h"
 #include "move.h"
 #include "tactics/util.h"
+#include "tactics/2lib.h"
 #include "pattern/spatial.h"
-#include "josekifix/josekifixload.h"
 #include "josekifix/josekifix.h"
+#include "josekifix/josekifixload.h"
 
+
+/**********************************************************************************************************/
+/* Override list stuff */
+
+/* Growable list of overrides.
+ * Ensures trailing null override always available (if zero'ed initially) */
+typedef struct {
+	override_t *overrides;
+	int          alloc;
+	int          len;
+} override_list_t;
+
+/* Growable list of overrides (2 overrides <and> check)
+ * Ensures trailing null override always available (if zero'ed initially) */
+typedef struct {
+	override2_t *overrides;
+	int          alloc;
+	int          len;
+} override2_list_t;
+
+/* Overrides and logged variations loaded at startup. */
+static override_list_t  joseki_overrides_list   = { 0, };
+static override2_list_t joseki_overrides2_list  = { 0, };	/* <and> checks */
+static override_list_t  logged_variations_list  = { 0, };
+static override2_list_t logged_variations2_list = { 0, };	/* <and> checks */
+
+
+/* Append override to list, realloc if necessary.  (override data is copied) */
+static void
+override_list_add(override_list_t *list, override_t *override)
+{
+	if (list->len + 2 > list->alloc) {  /* realloc */
+		size_t size = sizeof(*override);
+		int prev_n = list->alloc;
+		list->alloc *= 2;
+		list->alloc = MAX(list->alloc, 16);
+		int n = list->alloc;
+		list->overrides = crealloc(list->overrides, n * size);
+		memset(list->overrides + prev_n, 0, (n - prev_n) * size);   /* zero newly allocated memory ! */
+	}
+	
+	list->overrides[list->len++] = *override;
+}
+
+/* Append override to list, realloc if necessary.  (override data is copied) */
+static void
+override2_list_add(override2_list_t *list, override2_t *override)
+{
+	if (list->len + 2 > list->alloc) {  /* realloc */
+		size_t size = sizeof(*override);
+		int prev_n = list->alloc;
+		list->alloc *= 2;
+		list->alloc = MAX(list->alloc, 16);
+		int n = list->alloc;
+		list->overrides = crealloc(list->overrides, n * size);
+		memset(list->overrides + prev_n, 0, (n - prev_n) * size);   /* zero newly allocated memory ! */
+	}
+	
+	list->overrides[list->len++] = *override;
+}
+
+/**********************************************************************************************************/
+/* Override printing, comparing */
+
+static void
+print_ladder_check(char *idx, ladder_check_t *c)
+{
+	char *color = (c->own_color ? "own" : "other");
+	char *works = (c->works ? "" : "no");	
+	
+	fprintf(stderr, "  %sladder_%s%s = %s  [ ", works, color, idx, c->coord);
+	for (int i = 0; c->setup_own[i]; i++)
+		fprintf(stderr, "%s ", c->setup_own[i]);
+	fprintf(stderr, "]  [ ");
+	for (int i = 0; c->setup_other[i]; i++)
+		fprintf(stderr, "%s ", c->setup_other[i]);
+	fprintf(stderr, "]\n");
+}
+
+static void
+joseki_override_print(override_t *override, char *section)
+{
+	fprintf(stderr, "%s:\n", section);
+	fprintf(stderr, "  name = \"%s\"\n", override->name);
+	fprintf(stderr, "  prev = %s\n", override->prev);
+	fprintf(stderr, "  next = %s\n", override->next);
+	
+	if (override->coord_own)	fprintf(stderr, "  coord_own = %s\n", override->coord_own);
+	if (override->coord_other)	fprintf(stderr, "  coord_other = %s\n", override->coord_other);
+	if (override->coord_empty)	fprintf(stderr, "  coord_empty = %s\n", override->coord_empty);
+		
+
+	if (override->ladder_check.coord)  {  print_ladder_check("",  &override->ladder_check);  }
+	if (override->ladder_check2.coord) {  print_ladder_check("2", &override->ladder_check2);  }
+
+	fprintf(stderr, "  external_engine = [ ");
+	for (int q = 0; q < 4; q++)
+		fprintf(stderr, "%i ", override->external_engine_mode[q]);
+	fprintf(stderr, "]\n");
+
+	fprintf(stderr, "  hashes = { ");
+	for (int i = 0; i < 8; i++) {
+		fprintf(stderr, "0x%"PRIhash"%s ", override->hashes[i], (i != 7 ? "," : " }\n"));
+		if (i == 3)  fprintf(stderr, "\n             ");
+	}
+}
+
+/* like strcmp() but handle NULLs */
+static int
+same_str(char *s1, char *s2)
+{
+	return ((!s1 && !s2) ||
+		(s1 && s2 && !strcmp(s1, s2)));
+}
+
+static int
+ladder_check_cmp(ladder_check_t *c1, ladder_check_t *c2)
+{
+	if (!same_str(c1->coord, c2->coord) ||
+	    c1->works != c2->works ||
+	    c1->own_color != c2->own_color)
+		return 1;
+	
+	for (int i = 0; i < JOSEKIFIX_LADDER_SETUP_MAX; i++) {
+		if (!same_str(c1->setup_own[i], c2->setup_own[i]))
+			return 1;
+		if (!c1->setup_own[i] && !c2->setup_own[i])
+			break;
+	}
+	
+	for (int i = 0; i < JOSEKIFIX_LADDER_SETUP_MAX; i++) {
+		if (!same_str(c1->setup_other[i], c2->setup_other[i]))
+			return 1;
+		if (!c1->setup_other[i] && !c2->setup_other[i])
+			break;
+	}
+
+	return 0;
+}
+
+/* Compare 2 overrides (don't use that for sorting!) */
+static int
+override_cmp(override_t *o1, override_t *o2)
+{
+	return !(same_str(o1->prev, o2->prev) &&
+		 same_str(o1->next, o2->next) &&
+		 !memcmp(o1->hashes, o2->hashes, sizeof(o1->hashes)) &&
+		 same_str(o1->coord_own, o2->coord_own) &&
+		 same_str(o1->coord_other, o2->coord_other) &&
+		 same_str(o1->coord_empty, o2->coord_empty) &&
+		 !ladder_check_cmp(&o1->ladder_check, &o2->ladder_check) &&
+		 !ladder_check_cmp(&o1->ladder_check2, &o2->ladder_check2) &&
+		 !memcmp(o1->external_engine_mode, o2->external_engine_mode, sizeof(o1->external_engine_mode)));
+}
+
+
+/**********************************************************************************************************/
+/* Load from file */
+
+static void
+ladder_sanity_check(board_t *board, ladder_check_t *check, override_t *override)
+{
+	board_t b2;  board_copy(&b2, board);
+	board_t *b = &b2;
+
+	/* Check coords are valid */
+	
+	if (!valid_coord(check->coord)) {
+		board_print(board, stderr);	/* orig board, without setup stones */
+		die("josekifix: \"%s\": invalid ladder coord '%s', aborting. (run with -d5 to see previous moves)\n",
+		    override->name, check->coord);
+	}
+	
+	int n = JOSEKIFIX_LADDER_SETUP_MAX;
+	for (int i = 0; i < n && check->setup_own[i]; i++)
+		if (!valid_coord(check->setup_own[i])) {
+			board_print(board, stderr);
+			die("josekifix: \"%s\": invalid ladder setup_own coord '%s', aborting. (run with -d5 to see previous moves)\n",
+			    override->name, check->setup_own[i]);
+		}
+
+	for (int i = 0; i < n && check->setup_other[i]; i++)
+		if (!valid_coord(check->setup_other[i])) {
+			board_print(board, stderr);
+			die("josekifix: \"%s\": invalid ladder setup_other coord '%s', aborting. (run with -d5 to see previous moves)\n",
+			    override->name, check->setup_other[i]);
+		}
+	
+	/* Check board setup is sane */
+	
+	if (!josekifix_ladder_setup(b, 0, check)) {
+		board_print(board, stderr);
+		die("josekifix: \"%s\": bad ladder setup, some invalid move(s), aborting. (run with -d5 to see previous moves)\n",
+		    override->name);
+	}
+	
+	enum stone own_color = board_to_play(board);
+	enum stone ladder_color = (check->own_color ? own_color : stone_other(own_color));
+	coord_t c = str2coord(check->coord);
+	group_t g = board_get_2lib_neighbor(b, c, stone_other(ladder_color));
+
+	if (!g) {
+		board_print(board, stderr);	/* orig board */
+		board_print(b, stderr);		/* ladder setup board */
+		die("josekifix: \"%s\": bad ladder check, no ladder at %s ! aborting. (run with -d5 to see previous moves)\n",
+		    override->name, check->coord);
+	}
+
+	bool good_color = (board_at(b, g) == stone_other(ladder_color));
+	if (!good_color) {
+		board_print(board, stderr);	/* orig board */
+		board_print(b, stderr);		/* ladder setup board */
+		die("josekifix: \"%s\": ladder check at %s: wrong color, aborting. (run with -d5 to see previous moves)\n",
+		    override->name, check->coord);
+	}
+}
+
+/* Common sanity checks for [override] and [log] sections */
+static char*
+common_sanity_checks(board_t *b, override_t *override)
+{
+	if (!override->name || !override->name[0]) {
+		board_print(b, stderr);
+		die("josekifix: this override has no name, aborting. (run with -d5 to see previous moves)\n");
+	}
+
+	if (!valid_coord(override->prev) && strcmp(override->prev, "pass")) {
+		board_print(b, stderr);
+		die("josekifix: \"%s\": invalid prev move '%s', aborting. (run with -d5 to see previous moves)\n",
+		    override->name, override->prev);
+	}
+
+	if (!valid_coord(override->next) && strcmp(override->next, "pass")) {
+		board_print(b, stderr);
+		die("josekifix: \"%s\": invalid next move '%s', aborting. (run with -d5 to see previous moves)\n",
+		    override->name, override->next);
+	}
+
+	/* Already checked by add_override() but doesn't hurt ... */
+	char *around_str = NULL;
+	if (override->coord_own)	around_str = override->coord_own;
+	if (override->coord_other)	around_str = override->coord_other;
+	if (override->coord_empty)	around_str = override->coord_empty;
+	
+	if (around_str && !valid_coord(around_str)) {
+		board_print(b, stderr);
+		die("josekifix: \"%s\": invalid around coord '%s', aborting. (run with -d5 to see previous moves)\n",
+		    override->name, around_str);
+	}
+		
+	if (override->ladder_check.coord)   ladder_sanity_check(b, &override->ladder_check, override);
+	if (override->ladder_check2.coord)  ladder_sanity_check(b, &override->ladder_check2, override);
+	
+	/* Not checking hashes ... */
+
+	return around_str;
+}
+
+/* Check override is sane, help locate bad override otherwise. */
+static void
+override_sanity_checks(board_t *b, override_t *override)
+{
+	/* Common checks first */
+	char *around_str = common_sanity_checks(b, override);
+	
+	/* Warn if moves are too far apart. */
+	coord_t prev = str2coord(override->prev);
+	coord_t next = str2coord(override->next);
+	coord_t around = (around_str ? str2coord(around_str) : pass);
+
+#define point_dist(a, b)  (is_pass(a) || is_pass(b) ? 0 : roundf(coord_distance(a, b)))
+	
+	int dist = point_dist(prev, next);
+	if (dist > 8) {
+		board_print(b, stderr);
+		fprintf(stderr, "josekifix: \"%s\": big distance between prev move (%s) and next move (%s), bad override coords ?\n\n",
+			override->name, override->prev, override->next);
+	}
+
+	dist = point_dist(prev, around);
+	if (dist > 6) {
+		board_print(b, stderr);		
+		fprintf(stderr, "josekifix: \"%s\": big distance between prev move (%s) and around coord (%s), bad override coords ?\n\n",
+			override->name, override->prev, coord2sstr(around));
+	}
+
+	dist = point_dist(next, around);
+	if (dist > 6) {
+		board_print(b, stderr);		
+		fprintf(stderr, "josekifix: \"%s\": big distance between next move (%s) and around coord (%s), bad override coords ?\n\n",
+			override->name, override->next, coord2sstr(around));
+	}
+
+	// TODO if next move, check it's inside match pattern ...
+}
+
+/* Check log is sane, help locate bad override otherwise. */
+static void
+log_sanity_checks(board_t *b, override_t *override)
+{
+	/* Common checks first */
+	char *around_str = common_sanity_checks(b, override);
+
+	/* Warn if moves are too far apart.
+	 * (only check prev and around, logs have dummy next) */
+	coord_t prev = str2coord(override->prev);
+	coord_t around = (around_str ? str2coord(around_str) : pass);
+	
+	int dist = point_dist(prev, around);
+	if (dist > 6) {
+		board_print(b, stderr);
+		fprintf(stderr, "josekifix: \"%s\": big distance between prev move (%s) and around coord (%s), bad override coords ?\n\n",
+			override->name, override->prev, coord2sstr(around));
+	}
+}
+
+/* Add a new override to the set of checked overrides. */
+static void
+josekifix_add_override(board_t *b, override_t *override)
+{
+	/* Don't add duplicates */
+	for (int i = 0; i < joseki_overrides_list.len; i++)
+		if (!override_cmp(override, &joseki_overrides_list.overrides[i]))
+			return;	
+
+	override_sanity_checks(b, override);
+	override_list_add(&joseki_overrides_list, override);
+}
+
+/* Add new override and check (2 overrides) to the set of checked overrides. */
+static void
+josekifix_add_override_and(board_t *b, override_t *override1, override_t *override2)
+{
+	/* Don't add duplicates */
+	for (int i = 0; i < joseki_overrides2_list.len; i++)
+		if (!override_cmp(override1, &joseki_overrides2_list.overrides[i].override1) &&
+		    !override_cmp(override2, &joseki_overrides2_list.overrides[i].override2))
+			return;	
+
+	override_sanity_checks(b, override1);
+	/* Skip override2 sanity check (long distance warning but that's ok here). */
+	//override_sanity_checks(b, override2);
+	
+	override2_t and_check = { 0, };
+	and_check.override1 = *override1;
+	and_check.override2 = *override2;
+	
+	override2_list_add(&joseki_overrides2_list, &and_check);
+}
+
+/* Add a new logged variation to the set of logged variations.
+ * They work like overrides except they only affect logging :
+ * They don't interfere with game moves. */
+static void
+josekifix_add_logged_variation(board_t *b, override_t *log)
+{
+	// in this case override must have a next move (not pass),
+	// even though it's ignored.
+	log->next = "A1";
+	
+	// don't add duplicates
+	for (int i = 0; i < logged_variations_list.len; i++)
+		if (!override_cmp(log, &logged_variations_list.overrides[i]))
+			return;	
+
+	log_sanity_checks(b, log);
+	override_list_add(&logged_variations_list, log);
+}
+
+static void
+josekifix_add_logged_variation_and(board_t *b, override_t *log1, override_t *log2)
+{
+	// in this case overrides must have a next move (not pass),
+	// even though it's ignored.
+	log1->next = "A1";
+	log2->next = "A1";
+	
+	// don't add duplicates
+	for (int i = 0; i < logged_variations2_list.len; i++)
+		if (!override_cmp(log1, &logged_variations2_list.overrides[i].override1) &&
+		    !override_cmp(log2, &logged_variations2_list.overrides[i].override2))
+			return;	
+
+	log_sanity_checks(b, log1);
+	/* Skip override2 sanity check (long distance warning but that's ok here). */
+	//log_sanity_checks(b, log2);
+
+	override2_t and_check = { 0, };
+	and_check.override1 = *log1;
+	and_check.override2 = *log2;
+
+	override2_list_add(&logged_variations2_list, &and_check);
+}
+
+override_t  *joseki_overrides = NULL;
+override2_t *joseki_overrides2 = NULL;
+override_t  *logged_variations = NULL;
+override2_t *logged_variations2 = NULL;
+
+/* Load josekifix overrides from file.
+ * Debugging: to get a dump of all entries, run                      'pachi -d4'
+ *            to get a dump of all entries + earlier positions, run  'pachi -d5'  */
+bool
+josekifix_load(void)
+{
+	const char *fname = "josekifix.gtp";
+	FILE *f = fopen_data_file(fname, "r");
+	if (!f) {
+		if (DEBUGL(3))			perror(fname);
+		if (DEBUGL(2))			fprintf(stderr, "Joseki fixes: file %s missing\n", fname);
+		if (get_josekifix_required())	die("josekifix required but %s missing, aborting.\n", fname);
+		if (DEBUGL(2))			fprintf(stderr, "Joseki fixes disabled\n");
+		return false;
+	}
+	if (DEBUGL(2))  fprintf(stderr, "Loading joseki fixes ...\n");
+
+	DEBUG_QUIET();		// Turn off debugging (only want debug msg inside josekifixload engine)
+	board_t *b = board_new(19, NULL);
+	engine_t e;  engine_init(&e, E_JOSEKIFIXLOAD, NULL, NULL);
+	time_info_t ti[S_MAX];
+	ti[S_BLACK] = ti_none;
+	ti[S_WHITE] = ti_none;
+	char buf[4096];
+	gtp_t gtp;  gtp_init(&gtp, b);
+	for (int lineno = 1; fgets(buf, 4096, f); lineno++) {
+		/* Pachi version check */
+		if (str_prefix("# Pachi ", buf)) {
+			double wanted  = atof(buf + strlen("# Pachi "));
+			if (saved_debug_level > 3) fprintf(stderr, "checking version >= %.2f\n", wanted);
+			if (PACHI_VERNUM < wanted)
+				die("%s: need pachi version >= %.2f\n", fname, wanted);
+		}
+		
+		gtp.quiet = true;  // XXX fixme, refactor
+		enum parse_code c = gtp_parse(&gtp, b, &e, ti, buf);  /* quiet */
+		/* TODO check gtp command didn't gtp_error() also, will still return P_OK on error ... */
+		if (c != P_OK && c != P_ENGINE_RESET)
+			die("%s:%i  gtp command '%s' failed, aborting.\n", fname, lineno, buf);
+	}
+	engine_done(&e);
+	board_delete(&b);
+	DEBUG_QUIET_END();
+
+	if (DEBUGL(3))  fprintf(stderr, "Loaded %i overrides (and: %i), %i logged variations (and: %i)\n",
+				joseki_overrides_list.len, joseki_overrides2_list.len,
+				logged_variations_list.len, logged_variations2_list.len);
+	else if (DEBUGL(2))  fprintf(stderr, "Loaded %i overrides.\n", joseki_overrides_list.len);
+
+	joseki_overrides   = joseki_overrides_list.overrides;
+	joseki_overrides2  = joseki_overrides2_list.overrides;
+	logged_variations  = logged_variations_list.overrides;
+	logged_variations2 = logged_variations2_list.overrides;
+
+	fclose(f);
+	return true;
+}
+
+
+/**********************************************************************************************************/
+
+/* Fill in override hashes from board position (all rotations) */
+static void
+joseki_override_fill_hashes(override_t *override, board_t *b)
+{
+	enum stone color = last_move(b).color;	// last move color
+	
+	coord_t around = last_move(b).coord;
+	if (override->coord_empty)  around = str2coord(override->coord_empty);
+	if (override->coord_own)    around = str2coord(override->coord_own);
+	if (override->coord_other)  around = str2coord(override->coord_other);
+	
+	for (int rot = 0; rot < 8; rot++)
+		override->hashes[rot] = outer_spatial_hash_from_board_rot_d(b, around, color, rot, MAX_PATTERN_DIST);	
+}
+
+
+/**********************************************************************************************************/
+/* Parsing */
 
 #define board_captures(b)  (b->captures[S_BLACK] + b->captures[S_WHITE])
 
@@ -412,6 +893,10 @@ add_override(board_t *b, move_t *m, char *move_str)
 		else	      josekifix_add_logged_variation(b, &override);
 	}
 }
+
+
+/**********************************************************************************************************/
+/* JosekifixLoad engine */
 
 /* Record joseki overrides found in the GTP stream (extra data after play commands)
  * Debugging: to dump all overrides, run                      'pachi -d4'
